refactor(digitalocean): extract dns configuration to shared function
